rhomelさんからのレビューを元に変更を加えました。

+++ python/main.py
@@ -24,16 +24,21 @@ app.add_middleware(
 json_file="./items.json"
 
 def store_image(image: UploadFile = File(...)):
-    #image.readでcoroutine objectが生成されていたためbinary dataを取り出すためにasyncio.runを追加
+    # image.readでcoroutine objectが生成されていたためbinary dataを取り出すためにasyncio.runを追加
     image_bytes = asyncio.run(image.read())
     image_hash = hashlib.sha256(image_bytes).hexdigest()
     image_filename = f"{image_hash}.jpg"
 
-    with open(images / image_filename, "wb") as image_file:
-        image_file.write(image_bytes)
-
-    logger.info(f"Image saved: {image_filename}")
+    # imageファイルを複数回保存しないようにするため(sha256より同じ画像には同じファイル名が生成されることがsecureされている)
+    if os.path.isfile(image_filename):
+        logger.info(f"{image_filename} exists.")
+    else:
+        with open(images / image_filename, "wb") as image_file:
+            image_file.write(image_bytes)
+        logger.info(f"Image saved: {image_filename}")
+    
     return image_filename
+
 
 @app.get("/")
 def root():
@@ -49,19 +54,20 @@ def add_item(name: str = Form(...), category: str = Form(...), image: UploadFile
     
     image_filename = store_image(image)
     
-    if not {'name': name, 'category': category, 'image_name': image_filename} in items['items']:
-        items['items'].append({'name': name, 'category': category, 'image_filename': image_filename})
+    items['items'].append({'name': name, 'category': category, 'image_filename': image_filename})
 
     with open(json_file, mode='w') as j:
         json.dump(items, j)
         
     return {"message": f"item received: {name}, {category}, {image_filename}"}
 
+
 @app.get("/items")
 def get_items():
     with open(json_file, mode='r') as getfile:
         items = json.load(getfile)
     return items
+
 
 @app.get("/image/{image_name}")
 async def get_image(image_name):
@@ -76,6 +82,7 @@ async def get_image(image_name):
         image = images / "default.jpg"
 
     return FileResponse(image)
+
 
 @app.get("/items/{item_id}")
 def get_item(item_id: int):
